+++ src/components/gcs-search/searchControl.js
@@ -32,6 +32,8 @@ export default class SearchControl extends Control {
         if (this.searchString) {
             this.handleSearch({keyCode: 13, target: this.input, preventDefault: () => {
                 // noop
+            }, stopPropagation: () => {
+                // noop
             }});
         }
         super.setMap(map);
@@ -42,6 +44,7 @@ export default class SearchControl extends Control {
 
         if (e.keyCode === 13) {
             e.preventDefault();
+            e.stopPropagation();
             this.clearResults();
             if (this.search) {
                 this.search.search(elem.value).then((resp) => this.renderResponse(resp));
@@ -53,6 +56,8 @@ export default class SearchControl extends Control {
         if (property.key === "searchString") {
             this.input.value = this.get("searchString");
             this.handleSearch({keyCode: 13, target: this.input, preventDefault: () => {
+                // noop
+            }, stopPropagation: () => {
                 // noop
             }});
         }