+++ test/gdik-input/GDIKInput.spec.js
@@ -402,4 +402,24 @@ describe("public functions", () => {
 
         expect(gcsMapGetImage).toHaveBeenCalled();
     });
+
+    it("should call map fit method when component have a value", async () => {
+        // eslint-disable-next-line no-empty-function
+        const gcsMapFit = jest.spyOn(GCSMap.prototype, "fit").mockImplementation(() => {}),
+            component = new GDIKInput(),
+            geometry = {"type": "Point", "coordinates": [1, 1]},
+            jsonValue = {"type": "FeatureCollection", "features": [{"type": "Feature", "geometry": geometry}]},
+            value = JSON.stringify(jsonValue);
+
+        component.setAttribute("value", value);
+        component.setAttribute("draw-type", "Point");
+
+        await component.connectedCallback();
+
+        expect(component.value).toEqual(jsonValue);
+
+        component.centerToFeature();
+
+        expect(gcsMapFit).toHaveBeenCalledWith(geometry);
+    });
 });